test: move Review tests to specific folder
this commit moves Review tests to specific folder
